Move merge projections tests to under optimize projections
